+++ NEWS.md
@@ -7,6 +7,11 @@
 * `fp_text` gained arguments to specify different fonts when mixing 
   CJK and latin characters: `cs.family`, `eastasia.family`, `hansi.family`.
 
+## Issues
+
+* fix issue with document properties where values have to be html-escaped
+
+
 # officer 0.3.16
 
 ## New features